Potential fix for code scanning alert no. 4: Implicit narrowing conversion in compound assignment

+++ src/main/java/me/jellysquid/mods/sodium/client/gl/arena/staging/MappedStagingBuffer.java
@@ -97,7 +97,7 @@ public class MappedStagingBuffer implements StagingBuffer {
             commandList.flushMappedRange(this.mappedBuffer.map, this.start, this.pos - this.start);
         }
 
-        int bytes = 0;
+        long bytes = 0L;
 
         for (CopyCommand command : consolidateCopies(this.pendingCopies)) {
             bytes += command.bytes;
